+++ Nbm/src/test/resources/projForTest/src/completion/checkAfterDotFiltering.kt
@@ -1,0 +1,9 @@
+package completion
+
+fun topLevelNonExtension() {}
+fun String.stringExtension(): String = this
+
+fun testAfterDot() {
+    val s = "hello"
+    s.<caret>
+}